backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
paramiko>=3.4.0
python-dotenv>=1.0.0
websockets>=12.0
python-multipart>=0.0.6
